MyLab v0.1
We are doing a Google Glass project, where we aim to increase learning
for students by 1500% compared to normal lectures. This is done by
having a personal learning assistant with you during experiments. This
assistant will go through step by step instructions on how to perform a
certain task, in the pace you want. This concept can easily be scaled
up to include many different types of tutorials, from baking a cake, to
assembling that pesky IKEA sofa.